Add a very simple and rough telescope picker
This makes a lot of assumption, for instance:
* that zk is in the $PATH,
* that the `jsonl` format remains with one JSON object per line and that
this object holds the metadata for a zettel.

Also, I’m deliberately using the zk command when possible instead of the
LSP. That’s because so far the list command is more than enough and some
users may chose not to run the LSP (which can be separately configured
from https://github.com/neovim/nvim-lspconfig)

This unpolished version is hopefully the chance to get comments on where
we might be heading.